fix multiplayer race tracker
