+++ packages/app/src/app/race/_components/race/race-multiplayer.tsx
@@ -34,6 +34,7 @@ import type { Snippet } from "@prisma/client";
 import type { User } from "next-auth";
 import type { Socket } from "socket.io-client";
 import { RaceTimeStampProps, ReplayTimeStampProps } from "./types";
+import RaceTrackerMultiplayer from "./race-tracker-mutliplayer";
 
 type Participant = Omit<
   GameStateUpdatePayload["raceState"]["participants"][number],
@@ -470,11 +471,10 @@ export default function RaceMultiplayer({
           <>
             {raceId && code ? (
               participants.map((p) => (
-                <RaceTracker
+                <RaceTrackerMultiplayer
                   key={p.id}
                   position={p.position}
                   participantId={p.id}
-                  codeLength={code.length}
                 />
               ))
             ) : position && code ? (
